+++ snowflake/pkg/utils/timestamps/timestamps.go
@@ -20,18 +20,14 @@ import (
 	"time"
 )
 
-func ParseTimestamp(t string, now time.Time, defaultT ...time.Time) (string, error) {
-	defaultTimestamp := now
-	if len(defaultT) > 0 {
-		defaultTimestamp = defaultT[0]
-	}
+func ParseTimestamp(t string, now time.Time) (string, error) {
 	ts, err := func() (time.Time, error) {
 		fields := strings.Split(t, "-")
 		if len(fields) == 0 {
-			return defaultTimestamp, nil
+			return now, nil
 		}
 		if len(fields) > 1 && fields[0] != "now" {
-			return defaultTimestamp, fmt.Errorf("bad timestamp: %s", t)
+			return now, fmt.Errorf("bad timestamp: %s", t)
 		}
 		if len(fields) == 1 {
 			return now, nil
@@ -39,11 +35,11 @@ func ParseTimestamp(t string, now time.Time, defaultT ...time.Time) (string, err
 		if len(fields) == 2 {
 			d, err := time.ParseDuration(fields[1])
 			if err != nil {
-				return defaultTimestamp, fmt.Errorf("bad timestamp: %s", t)
+				return now, fmt.Errorf("bad timestamp: %s", t)
 			}
 			return now.Add(-d), nil
 		}
-		return defaultTimestamp, fmt.Errorf("bad timestamp: %s", t)
+		return now, fmt.Errorf("bad timestamp: %s", t)
 	}()
 	if err != nil {
 		return "", nil
